+++ x/epochs/handler.go
@@ -1,0 +1,19 @@
+package epochs
+
+import (
+	"fmt"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+
+	"github.com/tharsis/evmos/x/epochs/keeper"
+	"github.com/tharsis/evmos/x/epochs/types"
+)
+
+// NewHandler returns a handler for epochs module messages
+func NewHandler(k keeper.Keeper) sdk.Handler {
+	return func(_ sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
+		errMsg := fmt.Sprintf("unrecognized %s message type: %T", types.ModuleName, msg)
+		return nil, sdkerrors.Wrap(sdkerrors.ErrUnknownRequest, errMsg)
+	}
+}