+++ x/erc20/keeper/evm_hooks_test.go
@@ -6,6 +6,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
+
 	"github.com/tharsis/evmos/x/erc20/types"
 )
 
@@ -88,7 +89,7 @@ func (suite *KeeperTestSuite) TestEvmHooksRegisterERC20() {
 			balance := suite.app.BankKeeper.GetBalance(suite.ctx, sdk.AccAddress(suite.address.Bytes()), types.CreateDenom(contractAddr.String()))
 			suite.Commit()
 			if tc.result {
-				// Check if the execution was successfull
+				// Check if the execution was successful
 				suite.Require().Equal(int64(10), balance.Amount.Int64())
 			} else {
 				// Check that no changes were made to the account
@@ -151,7 +152,7 @@ func (suite *KeeperTestSuite) TestEvmHooksRegisterCoin() {
 			cosmosBalance = suite.app.BankKeeper.GetBalance(suite.ctx, sender, metadata.Base)
 
 			if tc.result {
-				// Check if the execution was successfull
+				// Check if the execution was successful
 				suite.Require().NoError(err)
 				suite.Require().Equal(cosmosBalance.Amount, sdk.NewInt(tc.mint-tc.burn+tc.reconvert))
 			} else {