+++ x/erc20/keeper/gov_hooks.go
@@ -5,6 +5,7 @@ import (
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+
 	"github.com/tharsis/evmos/x/erc20/types"
 )