+++ x/erc20/keeper/params.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/tharsis/evmos/x/erc20/types"
 )